+++ web/default/src/i18n/locales/en.json
@@ -1253,7 +1253,6 @@
     "DoubaoVideo": "DoubaoVideo",
     "Double check the configuration below. Your system will be locked until initialization is complete.": "Double check the configuration below. Your system will be locked until initialization is complete.",
     "Download": "Download",
-    "Draft": "Draft",
     "Draw": "Draw",
     "Drawing": "Drawing",
     "Drawing logs": "Drawing logs",
@@ -4415,7 +4414,6 @@
     "We could not load the setup status.": "We could not load the setup status.",
     "We will prompt your device to confirm using biometrics or your hardware key.": "We will prompt your device to confirm using biometrics or your hardware key.",
     "We'll be back online shortly.": "We'll be back online shortly.",
-    "Will be removed": "Will be removed",
     "Web search": "Web search",
     "Web Search": "Web Search",
     "Webhook Configuration:": "Webhook Configuration:",
