+++ web/default/src/i18n/locales/fr.json
@@ -1253,7 +1253,6 @@
     "DoubaoVideo": "DoubaoVideo",
     "Double check the configuration below. Your system will be locked until initialization is complete.": "Vérifiez la configuration ci-dessous. Votre système sera verrouillé jusqu'à ce que l'initialisation soit terminée.",
     "Download": "Télécharger",
-    "Draft": "Brouillon",
     "Draw": "Dessin",
     "Drawing": "Dessin",
     "Drawing logs": "Journaux de dessin",
@@ -4415,7 +4414,6 @@
     "We could not load the setup status.": "Nous n'avons pas pu charger l'état de la configuration.",
     "We will prompt your device to confirm using biometrics or your hardware key.": "Nous allons demander à votre appareil de confirmer en utilisant la biométrie ou votre clé matérielle.",
     "We'll be back online shortly.": "Nous serons de retour en ligne sous peu.",
-    "Will be removed": "Sera supprimé",
     "Web search": "Recherche web",
     "Web Search": "Recherche web",
     "Webhook Configuration:": "Configuration du Webhook :",
