+++ web/default/src/i18n/locales/ja.json
@@ -1253,7 +1253,6 @@
     "DoubaoVideo": "DoubaoVideo",
     "Double check the configuration below. Your system will be locked until initialization is complete.": "下記の設定を再確認してください。初期化が完了するまでシステムはロックされます。",
     "Download": "ダウンロード",
-    "Draft": "下書き",
     "Draw": "描画",
     "Drawing": "画像生成",
     "Drawing logs": "描画ログ",
@@ -4415,7 +4414,6 @@
     "We could not load the setup status.": "セットアップステータスを読み込めませんでした。",
     "We will prompt your device to confirm using biometrics or your hardware key.": "生体認証またはハードウェアキーを使用して確認するよう、デバイスにプロンプトが表示されます。",
     "We'll be back online shortly.": "まもなくオンラインに戻ります。",
-    "Will be removed": "削除予定",
     "Web search": "ウェブ検索",
     "Web Search": "Web 検索",
     "Webhook Configuration:": "Webhook設定:",
