+++ web/default/src/i18n/locales/ru.json
@@ -1253,7 +1253,6 @@
     "DoubaoVideo": "DoubaoVideo",
     "Double check the configuration below. Your system will be locked until initialization is complete.": "Дважды проверьте конфигурацию ниже. Ваша система будет заблокирована до завершения инициализации.",
     "Download": "Скачать",
-    "Draft": "Черновик",
     "Draw": "Рисование",
     "Drawing": "Рисование",
     "Drawing logs": "Журналы рисования",
@@ -4415,7 +4414,6 @@
     "We could not load the setup status.": "Не удалось загрузить статус настройки.",
     "We will prompt your device to confirm using biometrics or your hardware key.": "Мы предложим вашему устройству подтвердить действие с помощью биометрии или аппаратного ключа.",
     "We'll be back online shortly.": "Мы скоро вернемся в сеть.",
-    "Will be removed": "Будет удалено",
     "Web search": "Веб-поиск",
     "Web Search": "Веб-поиск",
     "Webhook Configuration:": "Конфигурация веб-хука:",
