+++ web/default/src/i18n/locales/vi.json
@@ -1253,7 +1253,6 @@
     "DoubaoVideo": "DoubaoVideo",
     "Double check the configuration below. Your system will be locked until initialization is complete.": "Kiểm tra kỹ lại cấu hình bên dưới. Hệ thống của bạn sẽ bị khóa cho đến khi quá trình khởi tạo hoàn tất.",
     "Download": "Tải xuống",
-    "Draft": "Bản nháp",
     "Draw": "Vẽ",
     "Drawing": "Vẽ",
     "Drawing logs": "Nhật ký vẽ",
@@ -4415,7 +4414,6 @@
     "We could not load the setup status.": "Chúng tôi không thể tải trạng thái thiết lập.",
     "We will prompt your device to confirm using biometrics or your hardware key.": "Chúng tôi sẽ yêu cầu thiết bị của bạn xác nhận bằng cách sử dụng sinh trắc học hoặc khóa bảo mật phần cứng của bạn.",
     "We'll be back online shortly.": "Chúng tôi sẽ sớm trực tuyến trở lại.",
-    "Will be removed": "Sẽ bị xóa",
     "Web search": "Tìm kiếm web",
     "Web Search": "Tìm kiếm web",
     "Webhook Configuration:": "Cấu hình Webhook:",
