+++ web/default/src/i18n/locales/zh.json
@@ -1253,7 +1253,6 @@
     "DoubaoVideo": "DoubaoVideo",
     "Double check the configuration below. Your system will be locked until initialization is complete.": "仔细检查以下配置。您的系统将在初始化完成前保持锁定状态。",
     "Download": "下载",
-    "Draft": "草稿",
     "Draw": "绘图",
     "Drawing": "绘图",
     "Drawing logs": "绘制日志",
@@ -4415,7 +4414,6 @@
     "We could not load the setup status.": "我们无法加载设置状态。",
     "We will prompt your device to confirm using biometrics or your hardware key.": "我们将提示您的设备使用生物识别或硬件密钥进行确认。",
     "We'll be back online shortly.": "我们将很快恢复在线。",
-    "Will be removed": "将被移除",
     "Web search": "网络搜索",
     "Web Search": "网页搜索",
     "Webhook Configuration:": "Webhook 配置：",
